Separate observation files
